consolidate test files
